api_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
requests
gunicorn
dj-database-url
django-cors-headers
django_filter